[42182444] Modify IsValid of all tickets to do local check (preReg) on transaction
- implemented ITxMemPoolTracker interface to track transactions added or removed from the memory pool.
- CTicketTxMemPoolTracker class to track mempool ticket transactions (add/remove):
    - handle notification processTransaction when a transaction added to the mempool
	- handle notification removeTx when a transaction removed from the mempool
    - cache maps for fast searching ticket transactions by id
	- get list of active txids in the mempool for the specific ticket id
- TxMemPool changes:
    - registration for objects that support ITxMemPoolTracker interface: minerPolicyEstimator and CTicketTxMemPoolTracker
	- batch_lookup method to get mempool transactions by the list of txids
    - some refactoring
- CPastelTicketMemPoolProcessor class for working with the specific mempool ticket transactions (same ticket id):
    - get the list of active mempool transactions hashes (txids) from the ticket mempool tracker object (CTicketTxMemPoolTracker)
	- get real ticket transactions from the mempool by txids (use new batch_lookup method with single lock)
	- parse ticket transactions and create tickets cache
	- various search methods in the cache:
	   - by primary key: TicketExists, FindTicket
       - by secondary key: TicketExistsBySecondaryKey, FindTicketBySecondaryKey
- Optimized creation of P2FMS transactions - removed unnecessary memory reallocations
- ChangeUserName ticket validation fixes:
    - pre-registration checks:
	   - throw runtime_error if username-change ticket transaction with the same username is already in mempool
	   - throw runtime_error if username-change ticket transaction with the same Pastel ID is already in mempool
	   - do not throw an error in this case: username is already registered in the blockchain (with PastelID2), but this name is no longer active -
	     username-change ticket transaction found in the mempool for the PastelID2
	- replay attack protection: do not throw runtime_error if username-change ticket exists in DB and block height or txid are different,
	  but that user has username-change ticket in the mempool
	- for regtest reduced number of blocks (10) to allow second username change
- google tests:
    - ported mempool tests from bitcoin_test to pastel_gtest
	- added tests for CTicketTxMemPoolTracker, CPastelTicketMemPoolProcessor, P2FMS changes
- python rpc-tests:
  - moved pastelid tests from mn_tickets.py to secure_container.py
  - moved and refactored all username-change ticket tests from mn_tickets.py to mn_ticket_username_change.py
  - added username-change ticket tests related to mempool validation:
      - ticket with the same username or same pastelid in the mempool (prereg)
	  - username already registered in blockchain with pastelid2, but not active - username-change ticket from pastelid2 exists in the mempool.
	    allow username change in this case
- created new group for python test scripts rpc-mnfast, added this group to CircleCI

+++ src/rpc/rawtransaction.cpp
@@ -495,7 +495,7 @@ UniValue createrawtransaction(const UniValue& params, bool fHelp)
         int64_t nLockTime = params[2].get_int64();
         if (nLockTime < 0 || nLockTime > std::numeric_limits<uint32_t>::max())
             throw JSONRPCError(RPC_INVALID_PARAMETER, "Invalid parameter, locktime out of range");
-        rawTx.nLockTime = nLockTime;
+        rawTx.nLockTime = static_cast<uint32_t>(nLockTime);
     }
     
     if (params.size() > 3 && !params[3].isNull()) {
@@ -510,7 +510,7 @@ UniValue createrawtransaction(const UniValue& params, bool fHelp)
                     strprintf("Invalid parameter, expiryheight should be at least %d to avoid transaction expiring soon",
                     nextBlockHeight + TX_EXPIRING_SOON_THRESHOLD));
             }
-            rawTx.nExpiryHeight = nExpiryHeight;
+            rawTx.nExpiryHeight = static_cast<uint32_t>(nExpiryHeight);
         } else {
             throw JSONRPCError(RPC_INVALID_PARAMETER, "Invalid parameter, expiryheight can only be used if Overwinter is active when the transaction is mined");
         }
